Move Sonar CFG UTs
