+++ analyzers/tests/SonarAnalyzer.UnitTest/CBDE/CheckMlirWithReference.cs
@@ -21,7 +21,7 @@
 using System.IO;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SonarAnalyzer.UnitTest.CFG;
+using SonarAnalyzer.UnitTest.CFG.Sonar;
 
 namespace SonarAnalyzer.UnitTest.CBDE
 {
