+++ analyzers/tests/SonarAnalyzer.UnitTest/LiveVariableAnalysis/LiveVariableAnalysisTest.cs
@@ -26,7 +26,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SonarAnalyzer.CFG.Sonar;
 using SonarAnalyzer.LiveVariableAnalysis;
 using SonarAnalyzer.LiveVariableAnalysis.CSharp;
-using SonarAnalyzer.UnitTest.CFG;
+using SonarAnalyzer.UnitTest.CFG.Sonar;
 using StyleCop.Analyzers.Lightup;
 
 namespace SonarAnalyzer.UnitTest.LiveVariableAnalysis
